+++ src/java/org/aspcfs/modules/troubletickets/actions/TroubleTicketsDocumentsFolders.java
@@ -51,7 +51,13 @@ public final class TroubleTicketsDocumentsFolders extends CFSModule {
       FileFolder thisFolder = (FileFolder) context.getFormBean();
       thisFolder.setParentId(context.getRequest().getParameter("parentId"));
       db = getConnection(context);
-      int ticketId = addTicket(context, db);
+      //Load the ticket and the organization
+      Ticket thisTicket = addTicket(context, db);
+      int ticketId = thisTicket.getId();
+      //Check access permission to organization record
+      if (!isRecordAccessPermitted(context, db, thisTicket.getOrgId())) {
+        return ("PermissionError");
+      }
       //Build array of folder trails
       ProjectManagementFileFolders.buildHierarchy(db, context);
       context.getRequest().setAttribute("fileFolder", thisFolder);
@@ -82,7 +88,13 @@ public final class TroubleTicketsDocumentsFolders extends CFSModule {
     boolean isValid = false;
     try {
       db = this.getConnection(context);
-      int ticketId = addTicket(context, db);
+      //Load the ticket and the organization
+      Ticket thisTicket = addTicket(context, db);
+      int ticketId = thisTicket.getId();
+      //Check access permission to organization record
+      if (!isRecordAccessPermitted(context, db, thisTicket.getOrgId())) {
+        return ("PermissionError");
+      }
       //Insert or update the folder
       FileFolder thisFolder = (FileFolder) context.getFormBean();
       boolean newFolder = (thisFolder.getId() == -1);
@@ -135,7 +147,13 @@ public final class TroubleTicketsDocumentsFolders extends CFSModule {
     Connection db = null;
     try {
       db = getConnection(context);
-      int ticketId = addTicket(context, db);
+      //Load the ticket and the organization
+      Ticket thisTicket = addTicket(context, db);
+      int ticketId = thisTicket.getId();
+      //Check access permission to organization record
+      if (!isRecordAccessPermitted(context, db, thisTicket.getOrgId())) {
+        return ("PermissionError");
+      }
       //Load the file folder
       FileFolder thisFolder = new FileFolder(db, Integer.parseInt(itemId));
       recordDeleted = thisFolder.delete(db);
@@ -174,7 +192,13 @@ public final class TroubleTicketsDocumentsFolders extends CFSModule {
     Connection db = null;
     try {
       db = getConnection(context);
-      int ticketId = addTicket(context, db);
+      //Load the ticket and the organization
+      Ticket thisTicket = addTicket(context, db);
+      int ticketId = thisTicket.getId();
+      //Check access permission to organization record
+      if (!isRecordAccessPermitted(context, db, thisTicket.getOrgId())) {
+        return ("PermissionError");
+      }
       //Load the file folder to be modified
       FileFolder thisFolder = (FileFolder) context.getFormBean();
       thisFolder.setId(Integer.parseInt(itemId));
@@ -205,7 +229,13 @@ public final class TroubleTicketsDocumentsFolders extends CFSModule {
     Connection db = null;
     try {
       db = getConnection(context);
-      int ticketId = addTicket(context, db);
+      //Load the ticket and the organization
+      Ticket thisTicket = addTicket(context, db);
+      int ticketId = thisTicket.getId();
+      //Check access permission to organization record
+      if (!isRecordAccessPermitted(context, db, thisTicket.getOrgId())) {
+        return ("PermissionError");
+      }
       //Load the folder
       FileFolder thisFolder = new FileFolder(db, Integer.parseInt(itemId));
       context.getRequest().setAttribute("FileFolder", thisFolder);
@@ -239,7 +269,13 @@ public final class TroubleTicketsDocumentsFolders extends CFSModule {
     Connection db = null;
     try {
       db = getConnection(context);
-      int ticketId = addTicket(context, db);
+      //Load the ticket and the organization
+      Ticket thisTicket = addTicket(context, db);
+      int ticketId = thisTicket.getId();
+      //Check access permission to organization record
+      if (!isRecordAccessPermitted(context, db, thisTicket.getOrgId())) {
+        return ("PermissionError");
+      }
       //Load the current folder
       FileFolder thisFolder = new FileFolder(db, Integer.parseInt(itemId));
       int folderId = Integer.parseInt(newFolderId);
@@ -279,7 +315,7 @@ public final class TroubleTicketsDocumentsFolders extends CFSModule {
    * @return Description of the Return Value
    * @throws SQLException Description of the Exception
    */
-  private int addTicket(ActionContext context, Connection db) throws SQLException {
+  private Ticket addTicket(ActionContext context, Connection db) throws SQLException {
     String ticketId = (String) context.getRequest().getParameter("tId");
     if (ticketId == null) {
       ticketId = (String) context.getRequest().getAttribute("tId");
@@ -294,15 +330,14 @@ public final class TroubleTicketsDocumentsFolders extends CFSModule {
    *
    * @param context  The feature to be added to the Ticket attribute
    * @param db       The feature to be added to the Ticket attribute
-   * @param ticketId The feature to be added to the Ticket attribute
    * @return Description of the Return Value
    * @throws SQLException Description of the Exception
    */
-  private int addTicket(ActionContext context, Connection db, String ticketId) throws SQLException {
+  private Ticket addTicket(ActionContext context, Connection db, String ticketId) throws SQLException {
     context.getRequest().setAttribute("tId", ticketId);
     Ticket thisTicket = new Ticket(db, Integer.parseInt(ticketId));
     context.getRequest().setAttribute("TicketDetails", thisTicket);
-    return thisTicket.getId();
+    return thisTicket;
   }
 }
